Updated Northwestern University stipend
Hi, I received an offer letter from the Northwestern physics PhD department about 2 weeks ago and the stipend is not the stipend listed here. It is $46,356 a year now, and the graduate students there are also unionized and have insurance that covers dental and vision (you might want to consider adding a column for those two things as well, because they are not so common).